LOTS of tweaks for programability in the future. Added InteractableObject. Changed Tree to CollidableObject. Collidable has a mobile bool value for puzzle solving. added a TriggerBox Class, unimplemented. Added a makeWalls function to the Mission class to aid in drawing in walls. draws straight lines of walls and adds them to the objects if you tell it where to start, how many pieces to place, and whether its horizontal or vertical.
Added ALL objects available to the Mission1. So very many lines of code. Mission 1 is a go for demo. Things are a bit off in correlation to the background image, but there isn't anything I can do for that without destroying the 64x64 grid. Anyway, pull it and check it out.